Pulling Functions proxy out to its own file
